requirements.txt
numpy
pandas
scikit-learn==1.3.2
matplotlib
seaborn
missingno
ipykernel
imbalanced-learn==0.11.0
xgboost
catboost
dagshub
pyarrow
joblib
dvc
dvc[s3]
mlflow<3
python-dotenv
pymongo
certifi
boto3
great_expectations
pyaml
#-e .